API

Modules

DensityInterface.DensityInterface

Types and constants

DensityInterface.LogFuncDensity

Functions and macros

DensityInterface.isdensitytype

DensityInterface.logdensityof

DensityInterface.logfuncdensity

Documentation

DensityInterface.DensityInterface — Module

DensityInterface

Trait-based interface for mathematical/statistical densities

DensityInterface.LogFuncDensity — Type

struct DensityInterface.LogFuncDensity{F}

Wraps a log-density function log_f to make it compatible with DensityInterface interface. Typically, LogFuncDensity(log_f) should not be called directly, logfuncdensity should be used instead.

DensityInterface.isdensitytype — Function

isdensitytype(type::Type)

Returns true if type is compatible with the DensityInterface interface, otherwise returns false by default.

DensityInterface.logdensityof — Function

logdensityof(density, x)::Real
logdensityof(density)

Computes the logarithmic value of density at a given point x, resp. returns a function that does so:

logy = logdensityof(some_density, x)
logdensityof(some_density, x) == logdensityof(some_density)(x)

and

log_f = logdensityof(density)
log_f(x) == logdensityof(density, x)
logfuncdensity(log_f) == density

logdensityof(density) defaults to Base.Fix1(logdensityof, density), but may be specialized for a density type. If so, logfuncdensity will typically have to be specialized for the return type of logdensityof as well.

The following identity must always hold:

logfuncdensity(logdensityof(density)) == density

DensityInterface.logfuncdensity — Function

logfuncdensity(log_f)

Returns a DensityInterface-compatible density that is defined by a given log-density function log_f:

density = logfuncdensity(log_f)
logdensityof(density, x) == log_f(x)
logdensityof(density) == log_f

logfuncdensity(log_f) returns an instance of DensityInterface.LogFuncDensity by default, but may be specialized to return something else depending on the type of log_f). If so, logdensityof will typically have to be specialized for the return type of logfuncdensity as well.

The following identity must always hold:

logfuncdensity(logdensityof(density)) == density