API

Modules

DensityInterface.DensityInterface

Types and constants

DensityInterface.LogDensityOf

DensityInterface.LogFuncDensity

Functions and macros

DensityInterface.isdensity

DensityInterface.logdensityof

DensityInterface.logfuncdensity

Documentation

DensityInterface.DensityInterface — Module

DensityInterface

Trait-based interface for mathematical/statistical densities

DensityInterface.LogDensityOf — Type

struct DensityInterface.LogDensityOf{D} <: Function

Computes the logarithmic value of density at given points

Typically, LogDensityOf(some_density) should not be called directly, logdensityof should be used instead.

DensityInterface.LogFuncDensity — Type

struct DensityInterface.LogFuncDensity{F<:Base.Callable}

Wraps a log-density function log_f to make it compatible with DensityInterface interface. Typically, LogFuncDensity(log_f) should not be called directly, logfuncdensity should be used instead.

DensityInterface.isdensity — Function

isdensity(obj::Any)

Returns true if obj is compatible with the DensityInterface interface, otherwise returns false by default.

DensityInterface.logdensityof — Function

logdensityof(density, x)::Real
logdensityof(density)::Function

Computes the logarithmic value of density at a given point x, resp. returns a function that does so:

logy = logdensityof(some_density, x)
logdensityof(some_density, x) == logdensityof(some_density)(x)

and

log_f = logdensityof(density)
log_f(x) == logdensityof(density, x)
logfuncdensity(log_f) == density

logdensityof(density) returns an instance of DensityInterface.LogDensityOf by default, but may be specialized to return something else, depending on the type of density). However, if a custom method is provided for logdensityof(density::SomeDensity), then a custom method must be provided for logfuncdensity as well that guarantees

logfuncdensity(logdensityof(density)) == density

DensityInterface.logfuncdensity — Function

logfuncdensity(log_f::Base.Callable)

Returns a DensityInterface-compatible density that is defined by a given log-density function fF:

density = logfuncdensity(log_f)
logdensityof(density, x) == log_f(x)
logdensityof(density) == log_f

logfuncdensity(log_f) returns an instance of DensityInterface.LogFuncDensity by default, but may be specialized to return something else depending on the type of log_f).