API

Interface

DensityInterface.hasdensity — Function

hasdensity(d)::Bool

Return true if d is compatible with the DensityInterface interface.

hasdensity(d) == true implies that d is either a density itself or has an associated density, e.g. a probability density function or a Radon–Nikodym derivative with respect to an implicit base measure. It also implies that the value of that density at given points can be calculated via logdensityof and densityof. ```

DensityInterface.logdensityof — Function

logdensityof(d, x)::Real

Compute the logarithmic value of density d or it's associated density at a given point x.

julia> hasdensity(d)
true

julia> logy = logdensityof(d, x); logy isa Real
true

See also hasdensity and densityof.

DensityInterface.logdensityof — Method

logdensityof(d)

Return a function that computes the logarithmic value of density d or its associated density at a given point.

julia> log_f = logdensityof(d); log_f isa Function
true

julia> log_f(x) == logdensityof(d, x)
true

logdensityof(d) defaults to Base.Fix1(logdensityof, d), but may be specialized. If so, logfuncdensity will typically have to be specialized for the return type of logdensityof as well.

logfuncdensity is the inverse of logdensityof, so logfuncdensity(log_f) must be equivalent to d.

DensityInterface.logfuncdensity — Function

logfuncdensity(log_f)

Return a DensityInterface-compatible density that is defined by a given log-density function log_f:

julia> d = logfuncdensity(log_f);

julia> hasdensity(d) == true
true

julia> logdensityof(d, x) == log_f(x)
true

logfuncdensity(log_f) returns an instance of DensityInterface.LogFuncDensity by default, but may be specialized to return something else depending on the type of log_f). If so, logdensityof will typically have to be specialized for the return type of logfuncdensity as well.

logfuncdensity is the inverse of logdensityof, so the following must hold true:

logfuncdensity(logdensityof(d)) is equivalent to d

logdensityof(logfuncdensity(log_f)) is equivalent to log_f.

See also hasdensity.

DensityInterface.densityof — Function

densityof(d, x)::Real

Compute the value of density d or its associated density at a given point x.

julia> hasdensity(d)
true

julia> densityof(d, x) == exp(logdensityof(d, x))
true

densityof(d, x) defaults to exp(logdensityof(d, x)), but may be specialized.

See also hasdensity and logdensityof.

densityof(d)

Return a function that computes the value of density d or its associated density at a given point.

julia> f = densityof(d);

julia> f(x) == densityof(d, x)
true

densityof(d) defaults to Base.Fix1(densityof, d), but may be specialized.

DensityInterface.densityof — Method

densityof(d)

Return a function that computes the value of density d or its associated density at a given point.

julia> f = densityof(d);

julia> f(x) == densityof(d, x)
true

densityof(d) defaults to Base.Fix1(densityof, d), but may be specialized.

Types

DensityInterface.LogFuncDensity — Type

struct DensityInterface.LogFuncDensity{F}

Wraps a log-density function log_f to make it compatible with DensityInterface interface. Typically, LogFuncDensity(log_f) should not be called directly, logfuncdensity should be used instead.

Test utility

DensityInterface.test_density_interface — Function

DensityInterface.test_density_interface(d, x, ref_logd_at_x; kwargs...)

Test if d is compatible with DensityInterface.

Tests that logdensityof(d, x) equals ref_logd_at_x and that the behavior of logdensityof(d), densityof(d, x) and densityof(d) is consistent.

Also tests if logfuncdensity(logdensityof(d)) returns a density equivalent to d in respect to the functions above.

The results of logdensityof(d, x) and densityof(d, x) are compared to ref_logd_at_x and exp(ref_logd_at_x) using isapprox. kwargs... are forwarded to isapprox.